T3C-886: Refactor CruxDisplay and AboutPage to fix CodeScene regressions
Extract hooks and components from CruxDisplay to reduce cyclomatic complexity:
- useTextTruncation: text truncation measurement logic
- useParsedSpeakers: speaker parsing memoization
- useCruxNavigation: navigation handlers
- CruxHoverContent: hover card UI component
- createExplanationCleaner: text cleaning utility

Extract AboutPage sections to reduce file size:
- FAQSections: FAQ and Report Creation FAQ components
- PrivacySecurity: privacy policy section

Cleanup:
- Remove unused useCallback import
- Deduplicate ContentGroup via export/import pattern
- Fix hook ordering (remove biome-ignore comments)

Results:
- Subtopic.tsx: CC 17 → Score 10.00
- about/page.tsx: 409 → 248 lines, Score 9.27
