+++ next-client/src/app/about/components/ContentGroup.tsx
@@ -1,0 +1,5 @@
+import { Col } from "@/components/layout";
+
+export const ContentGroup = ({ children }: React.PropsWithChildren) => (
+  <Col gap={3}>{children}</Col>
+);
